+++ resources/views/widgets/header.blade.php
@@ -1,0 +1,10 @@
+<?php $navbar = new \Illuminate\Support\Fluent([
+    'id'    => 'control',
+    'title' => 'Control',
+    'url'   => handles('orchestra::control'),
+    'menu'  => view('orchestra/control::widgets._menu'),
+]); ?>
+
+@decorator('navbar', $navbar)
+
+<br>