+++ src/Contracts/Command/Synchronizer.php
@@ -1,0 +1,11 @@
+<?php namespace Orchestra\Control\Contracts\Command;
+
+interface Synchronizer
+{
+    /**
+     * Re-sync administrator access control.
+     *
+     * @return void
+     */
+    public function handle();
+}